Ahmad & Anfal

+++ restaurants/views.py
@@ -59,6 +59,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect("signin")
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -74,7 +76,10 @@ def restaurant_create(request):
 
 def item_create(request, restaurant_id):
     form = ItemForm()
+
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or (request.user == restaurant.owner)) :
+        return redirect('404')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +94,13 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+    if not (request.user.is_staff or (request.user ==  restaurant_obj.owner )) :
+        return redirect('404')
+
+
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +114,15 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if  not request.user.is_staff :
+        return redirect('404')
+
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+
+def NO(request):
+    
+        return render(request,'404.html')
